==> notebooks/1_Preprocessing-SourceTableCleaning-FilteringandFormatting.ipynb <==
# 1. MAF Source Data Pre-Processing 
##   Filtering and Formatting

Version Updated Dec 11, Ready for publish

%matplotlib inline
from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"
from IPython.display import Markdown, display
def printmd(string):
    display(Markdown(string))
def printmd(string, color=None):
    colorstr = "<span style='color:{}'>{}</span>".format(color, string)
    display(Markdown(colorstr))
from __future__ import print_function
import numpy as np
import pandas as pd
from pandas import DataFrame   
from pandas import pivot
from pandas import tools
import matplotlib.pyplot as plt
from sklearn import preprocessing
import os
from os.path import join, dirname
from tqdm import tqdm_notebook
import cv2
import re
import io
import glob
import random
import warnings
import pickle
import cloudpickle
import sys 
import azureml
import string
from scipy import stats
import pip
import csv
import pyparsing
import tempfile  
import logging
from IPython.display import SVG
import mkl
import graphviz
from azure.storage.blob import BlockBlobService
import datetime

os.chdir('D://MAFData')
base_dir = 'D://MAFData'
WorkingDir = os.path.join(base_dir, 'Workingdir')

os.getcwd()

## Import Portfolio Source Data

blob_account_name = "<youraccountnamehere>" # fill in your blob account name
blob_account_key = "<youraccountkeyhere>"  # fill in your blob account key

mycontainer = "<yourcontainernamehere>"       # fill in the container name
myblobname = "<yourfile.whateverextension>"        # fill in the blob name.  in our case we had .rpt sql report files.
mydatafile = "Portfolio"        # fill in the output file name

#Read rpt from blob
mydata = pd.read_csv(Portfolio, sep='\t',error_bad_lines=False, index_col=False, dtype='unicode')
print(mydata.shape)

#Read in from rpt file from local drive for iterative work
# mydata = pd.read_csv('df_Portfolio.rpt', sep='\t',error_bad_lines=False, index_col=False, dtype='unicode')
# mydata.shape
# mydata.head(2)

#Account number is anonymized
mydata.shape
mydata.head(1)

#Filter to just the columns of interest

Portfolio = mydata[['Account_Number',
 'MISDATE',
 'ATTRITION_FLAG',
 'ACCOUNT_CREATION_DATE',
 'ACTUAL_CURRENT_BALANCE',
 'Credit_Limit',
 'LAST_ACTIVITY_DATE',
 'Last_Statement_Closing_Balance',
 'Last_Statement_Date',
 'Last_Statement_Minimum_Due',
 'Monthly_Salary',
 'Months_On_Book',
 'Next_Statement_Date',
 'Monthly_Salary',
 'Nationality_Group',
 'Next_Statement_Date',
 'No_Of_Due_Stmt',
 'Product',
 'APR_Rate',
 'Product_Fee_Type',
 'Product_Category']]

del mydata

#See revised dimensions
Portfolio.shape

#Reformat date, filter and review head and tail of dataframe.
Portfolio['LAST_ACTIVITY_DATE'] = Portfolio['LAST_ACTIVITY_DATE'].astype('datetime64[ns]')
Portfolio = Portfolio.sort_values(['Account_Number', 'MISDATE'], ascending=[True, True])
Portfolio = Portfolio.reset_index(drop=True)
Portfolio.shape
Portfolio.head(2)
Portfolio.tail(1)
list(Portfolio.columns.values)

#Drop garbage first column
Portfolio = Portfolio.drop(Portfolio.index[0])

#Fill in numerical NaN with '0'
Portfolio = Portfolio.reset_index(drop=True)
Portfolio = Portfolio.dropna(subset=['Account_Number'])
Portfolio = Portfolio.dropna(subset=['MISDATE'])
Portfolio['Last_Statement_Closing_Balance'].fillna(0, inplace=True)
Portfolio['ACTUAL_CURRENT_BALANCE'].fillna(0, inplace=True)
Portfolio['Last_Statement_Minimum_Due'].fillna(0, inplace=True)
Portfolio = Portfolio.reset_index(drop=True)
Portfolio.shape
Portfolio.head(1)

#Write out to pickle file for later reference
#output = open('Porfolio_Source_v2.pkl', 'wb')
#pickle.dump(Portfolio, output)
#output.close()

#### Filter Portfolio file to just 'open' or 'voluntary

Portfolio['ATTRITION_FLAG'].unique()
Portfolio = Portfolio.loc[(Portfolio['ATTRITION_FLAG']=="OPEN") | (Portfolio['ATTRITION_FLAG']=="VOLUNTARY"), :]
Portfolio['ATTRITION_FLAG'].unique()

# Fill in Numerical NaN with 0
Portfolio['Last_Statement_Closing_Balance'].fillna(0, inplace=True)
Portfolio['ACTUAL_CURRENT_BALANCE'].fillna(0, inplace=True)
Portfolio['Last_Statement_Minimum_Due'].fillna(0, inplace=True)
Portfolio['ATTRITION_FLAG'].unique()
Portfolio['ATTRITION_FLAG'] = Portfolio.ATTRITION_FLAG.astype(str)

Portfolio = Portfolio.reset_index(drop=True)
Portfolio.shape

#Format data types 
Portfolio = pd.DataFrame(Portfolio)
Portfolio['MISDATE'] = Portfolio['MISDATE'].astype('datetime64[ns]')
Portfolio['Account_Number'] = Portfolio['Account_Number'].astype('str')
Portfolio.shape


Portfolio.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False], inplace=True)
Portfolio.reset_index(drop=True, inplace=True)

### Take accounts where `Months_On_Book` > 6

Portfolio['Months_On_Book'] = Portfolio['Months_On_Book'].astype(int)
justaccounts = Portfolio.groupby('Account_Number').head(1)
justaccounts.shape

newaccounts = justaccounts[justaccounts['Months_On_Book']<7]
newaccounts.head(1)
newaccounts.shape
newaccounts = newaccounts['Account_Number']
type(newaccounts)

Portfolio['Account_Number'].nunique()
Portfolio = Portfolio[~Portfolio['Account_Number'].isin(newaccounts)]
Portfolio['Account_Number'].nunique()
Portfolio.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False], inplace=True)
Portfolio = Portfolio.reset_index(drop=True)
Portfolio.shape
Portfolio.head(1)
Portfolio.tail(1)

### Trim to 365 most recent timesteps for each account number


Portfolio = Portfolio.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False])
Portfolio = Portfolio.reset_index(drop=True)
Portfolio = Portfolio.groupby('Account_Number').head(365)
Portfolio = Portfolio.reset_index(drop=True)
Portfolio.shape
Portfolio.head(1)
Portfolio.tail(1)
#check whether Portfolio data frame has complete #365 day series
Portfolio.shape[0]/365

### Filter to those accounts with at least one year of account history, or 365 daily timesteps

Portfolio = Portfolio.groupby('Account_Number').filter(lambda x: len(x) > 364)
Portfolio = Portfolio.reset_index(drop=True)
Portfolio.shape
Portfolio.head(2)

#Confirm that our 365 timesteps are fully populated.
Portfolio.shape[0]/365


Portfolio.shape

### Format Dates and Create Date Derived Variables

Portfolio1 = Portfolio
#Format data types 
Portfolio1['Last_Statement_Date'] = Portfolio1['Last_Statement_Date'].astype('datetime64[ns]')
Portfolio1['Next_Statement_Date'] = Portfolio1['Next_Statement_Date'].astype('datetime64[ns]')
Portfolio1['ACCOUNT_CREATION_DATE']=pd.to_datetime(Portfolio1['ACCOUNT_CREATION_DATE'],format='%Y-%m-%d')
Portfolio1['ACCOUNT_CREATION_DATE'] = Portfolio1['ACCOUNT_CREATION_DATE'].astype('datetime64[ns]')
Portfolio1['Account_Number'] = Portfolio1['Account_Number'].astype(str)
Portfolio1['ATTRITION_FLAG'] = Portfolio1['ATTRITION_FLAG'].astype(str)
Portfolio1['LAST_ACTIVITY_DATE']=pd.to_datetime( Portfolio1['LAST_ACTIVITY_DATE'],format='%Y-%m-%d')
Portfolio1['LAST_ACTIVITY_DATE'].fillna(Portfolio1['ACCOUNT_CREATION_DATE'], inplace=True)

Portfolio1['Last_Statement_Minimum_Due'] = Portfolio1['Last_Statement_Minimum_Due'].replace(-np.inf, 0)
Portfolio1['Last_Statement_Minimum_Due'] = Portfolio1['Last_Statement_Minimum_Due'].replace(-np.nan, 0)
Portfolio1['Last_Statement_Minimum_Due'] = Portfolio1['Last_Statement_Minimum_Due'].replace(np.inf, 0)
Portfolio1['Last_Statement_Minimum_Due'] = Portfolio1['Last_Statement_Minimum_Due'].replace(np.nan, 0)

Portfolio1['Monthly_Salary'] = Portfolio1['Monthly_Salary'].astype('int')
Portfolio1['Credit_Limit'] = Portfolio1['Credit_Limit'].astype('int')
Portfolio1['Last_Statement_Closing_Balance'] = Portfolio1['Last_Statement_Closing_Balance'].astype('int')
Portfolio1['Last_Statement_Minimum_Due'] = Portfolio1['Last_Statement_Minimum_Due'].astype('int')

#Days on Book
Portfolio1['ACCOUNT_CREATION_DATE'] = pd.to_datetime(Portfolio1['ACCOUNT_CREATION_DATE'], format='%Y-%m-%d')
Portfolio1['Days_On_Book'] = Portfolio1['MISDATE'] - Portfolio1['ACCOUNT_CREATION_DATE']
Portfolio1['Days_On_Book'] = Portfolio1['Days_On_Book'].dt.days
Portfolio1['Days_On_Book'] = Portfolio1['Days_On_Book'].astype(int)

#Days from Last Statement
Portfolio1['Last_Statement_Date'] = pd.to_datetime(Portfolio1['Last_Statement_Date'], format='%Y-%m-%d')
Portfolio1['Days_From_Statement'] = Portfolio1['MISDATE']-Portfolio1['Last_Statement_Date']
Portfolio1['Days_From_Statement'] = Portfolio1['Days_From_Statement'].dt.days
Portfolio1['Days_From_Statement'] = Portfolio1['Days_From_Statement'].astype(int)

#Days from Last Activity
Portfolio1['LAST_ACTIVITY_DATE'].fillna(Portfolio1['Last_Statement_Date'], inplace=True)
Portfolio1['LAST_ACTIVITY_DATE'].fillna(Portfolio1['MISDATE'], inplace=True)
Portfolio1['LAST_ACTIVITY_DATE']=pd.to_datetime( Portfolio1['LAST_ACTIVITY_DATE'],format='%Y-%m-%d')
Portfolio1['Days_from_Last_Activity'] = Portfolio1['MISDATE'] - Portfolio1['LAST_ACTIVITY_DATE']
Portfolio1['Days_from_Last_Activity'] = Portfolio1['Days_from_Last_Activity'].dt.days
Portfolio1['Days_from_Last_Activity'] = Portfolio1['Days_from_Last_Activity'].replace(-np.inf, 0)
Portfolio1['Days_from_Last_Activity'] = Portfolio1['Days_from_Last_Activity'].replace(-np.nan, 0)
Portfolio1['Days_from_Last_Activity'] = Portfolio1['Days_from_Last_Activity'].astype(int)

Portfolio1.shape

Portfolio1.head(1)

#Reset index, order, check dimensions
Portfolio1 = Portfolio1.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False])
Portfolio1 = Portfolio1.reset_index(drop=True)

#Check counts of unique ids.

uniqueids =Portfolio1.groupby('Account_Number').head(1)
uniqueids = uniqueids.reset_index(drop=True)
uniqueids['ATTRITION_FLAG'].value_counts()

#### Write out interim file

output = open('Portfolio_cleaned.pkl', 'wb')
pickle.dump(Portfolio1, output)
output.close()

del Portfolio

## Import Statement Data from SQL Pull, Pickled and Stored in Blob

mycontainer = "<yourcontainername>"       # fill in the container name
myblobname = "<yourblobname"        # fill in the blob name
mydatafile = "Statements"        # fill in the output file name that you want

dirname = base_dir
blob_service = BlockBlobService(account_name=blob_account_name,
                 account_key=blob_account_key)
blob_service.get_blob_to_path(mycontainer, myblobname, mydatafile)


mydata = pd.read_csv(mydatafile, sep='\t',error_bad_lines=False, index_col=False, dtype='unicode')
print(mydata.shape)

Statements = mydata[['ACCOUNT_NUMBER', 'STATEMENT_DATE', 'BEGINNING_AR', 'CLOSING_AR',
       'TOTAL_MINIMUM_DUE', 'PAYMENT_DUE_DATE', 'POINTS_BALANCE_BEGINNING',
       'POINTS_EARNED', 'POINTS_REDEEMED', 'POINTS_BALANCE_CLOSING',
       'POINTS_ADJUSTED', 'STATEMENT_FROM_DATE', 'STATEMENT_TO_DATE',
       'STATEMENT_BALANCE', 'INSTALLMENTS_REMAINING', 'NEW_INSTALLMENT_FLAG',
       'BT_FLAG', 'LOP_FLAG', 'EPP_FLAG', 'AMF_FLAG']]

del mydata

#### Statements - Reduce and Reformat

Statements.head(1)

Statements = pd.DataFrame(Statements)
Statements = Statements.dropna(subset=['ACCOUNT_NUMBER'])
Statements = Statements.dropna(subset=['STATEMENT_DATE'])
Statements_Index = Statements.reset_index(drop=True)

Statements_Index = Statements_Index.sort_values(['ACCOUNT_NUMBER', 'STATEMENT_DATE'], ascending=[True, True])
Statements_Index = Statements_Index.reset_index(drop=True)
Statements_Index = Statements_Index[['ACCOUNT_NUMBER','STATEMENT_DATE','STATEMENT_FROM_DATE','STATEMENT_TO_DATE','STATEMENT_BALANCE','POINTS_REDEEMED','POINTS_BALANCE_CLOSING','POINTS_EARNED']]
Statements_Index.columns = ['Account_Number','STATEMENT_DATE','STATEMENT_FROM_DATE','STATEMENT_TO_DATE','STATEMENT_BALANCE','POINTS_REDEEMED','POINTS_BALANCE_CLOSING', 'POINTS_EARNED']
Statements_Index['STATEMENT_DATE'] = Statements_Index['STATEMENT_DATE'].astype('datetime64[ns]')
Statements_Index['STATEMENT_FROM_DATE'] = Statements_Index['STATEMENT_FROM_DATE'].astype('datetime64[ns]')
Statements_Index['STATEMENT_TO_DATE'] = Statements_Index['STATEMENT_TO_DATE'].astype('datetime64[ns]')
Statements_Index['Account_Number'] = Statements_Index['Account_Number'].astype('str')
Statements_Index['STATEMENT_BALANCE'] = Statements_Index['STATEMENT_BALANCE'].astype('float')
Statements_Index['STATEMENT_BALANCE'] = round(Statements_Index['STATEMENT_BALANCE'],0)
Statements_Index['STATEMENT_BALANCE'] = Statements_Index['STATEMENT_BALANCE'].astype('int')
Statements_Index['POINTS_EARNED'] = Statements_Index['POINTS_EARNED'].astype('int')
Statements_Index['POINTS_REDEEMED'] = Statements_Index['POINTS_REDEEMED'].astype('int')
Statements_Index['POINTS_BALANCE_CLOSING'] = Statements_Index['POINTS_BALANCE_CLOSING'].astype('int')

Statements_Index.head(2)
Statements_Index.shape

Statements_Index.dtypes

del Statements

### Write out Cleaned Statements DataFrame

Statements_Index.to_csv('Statements_cleaned.csv',sep=',',encoding='utf-8')

del Statements_Index

## Import CRM Data from SQL Pull Stored in Blob

mycontainer = "<YourContainerName>"       # fill in the container name
myblobname = "<YourBlobName>"        # fill in the blob name
mydatafile = "<YourDataName>"        # fill in the output file name

blob_service = BlockBlobService(account_name=blob_account_name,
                 account_key=blob_account_key)
blob_service.get_blob_to_path(mycontainer, myblobname, mydatafile)

CRM = pd.read_csv(mydatafile, sep='\t',error_bad_lines=False, index_col=False, dtype='unicode')
print(CRM.shape)

CRM.dtypes

#### Reformat Variables

CRM = CRM.sort_values(['Account_Number'], ascending=[ True])
CRM = CRM.reset_index(drop=True)
CRM = CRM[['Account_Number','Incident_Category','Incident_Type','Created_On']]
CRM['Incident_Category'] = CRM['Incident_Category'].astype('str')
CRM['Incident_Type'] = CRM['Incident_Type'].astype('str')
CRM['Created_On'] = CRM['Created_On'].astype('datetime64[ns]')
CRM = CRM.dropna(subset=['Account_Number'])
CRM = CRM.dropna(subset=['Created_On'])
CRM = CRM.reset_index(drop=True)

CRM.head(1)
CRM.shape
CRM.dtypes

#### Remove Very infrequent incident Types

print('number of unique account numbers',len(CRM.Account_Number.unique()))
print('number of unique incident categorices',(len(CRM.Incident_Category.unique())))
#CRM['Incident_Type'].value_counts()

CRM.shape
#remove incidents with less than 100 occurences
threshold = 99
counts = CRM['Incident_Type'].value_counts()

CRM2 = CRM.loc[CRM['Incident_Type'].isin(counts[counts > threshold].index), :]
CRM2 = CRM2.loc[(CRM2['Incident_Type']!="Inbound SMS")]
CRM2 = CRM2.reset_index(drop=True)
print('number of unique account numbers',len(CRM2.Account_Number.unique()))
print('number of unique incident categorices',(len(CRM2.Incident_Category.unique())))
CRM2.shape

CRM2.to_csv('CRM_cleaned.csv',sep=',',encoding='utf-8')

## Import Aggregated Transaction Data from SQL Pull Stored in Blob

mycontainer = "<yourcontainer>"       # fill in the container name
myblobname = "<yourblobname>"        # fill in the blob name
mydatafile = "TransactionsDailyAgg"        # fill in the output file name

blob_service = BlockBlobService(account_name=blob_account_name,
                 account_key=blob_account_key)
blob_service.get_blob_to_path(mycontainer, myblobname, mydatafile)

Transactions_DailyAgg = pd.read_csv(mydatafile, sep='\t',error_bad_lines=False, index_col=False, dtype='unicode')
print(Transactions_DailyAgg.shape)


Transactions_DailyAgg.dtypes

Transactions_DailyAgg.rename(columns = {
      'sum' : 'Billing_Amount_Sum',
      'count' : 'Transaction_Count'},inplace = True)

output = open('Transactions_cleaned.pkl', 'wb')
pickle.dump(Transactions_DailyAgg, output)
output.close()

#Transactions_DailyAgg.to_csv('Transactions_DailyAgg_cleaned.csv',sep=',',encoding='utf-8')

### End Part 1 - Cleaning each source table
