==> notebooks/2_Preprocessing-SourceTables-Merging-and-Folding.ipynb <==
## 2. Source Data Merging and Folding

Last Updated: December 11, 2018

%matplotlib inline

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

from __future__ import print_function
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
from IPython.display import SVG
from sklearn import preprocessing
import os
from os.path import join, dirname
from os.path import dirname, join
from tqdm import tqdm_notebook
import pandas as pd
from pandas import pivot
from pandas import tools
import numpy as np
import cv2
import re
import io
import sys 
import pip
import csv
from sklearn.model_selection import train_test_split
import glob
import random
import warnings
from pandas import DataFrame   
import pickle
import cloudpickle
import azureml
import string
from scipy import stats
import pyparsing
import tempfile  
import logging
import graphviz

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

### Working Directory Definition

os.chdir('D://MAFData/v2data')
base_dir = 'D://MAFData/v2data'
WorkingDir = os.path.join(base_dir, 'Workingdir')

print(base_dir)
os.getcwd()

### Read in Cleaned Source Data

import pprint, pickle
PATH = 'D://MAFData//'
os.chdir('D://MAFData')
os.getcwd()

import datetime
t1 = datetime.datetime.now(); print(t1)

with open(PATH + 'Portfolio_cleaned.pkl', 'rb') as f:
    Portfolio_cleaned = pickle.load(f)

t2 = datetime.datetime.now(); print(t2) 
time_spent = t2-t1; print(time_spent)

Portfolio_cleaned.shape[0]/365
Portfolio_cleaned.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False], inplace=True)
Portfolio_cleaned.reset_index(drop=True, inplace=True)

### Count of 'ATTRITION_FLAG' == 'VOLUNTARY'

    extract Account_Number of VOLUNTARY



Portfolio_cleaned = Portfolio_cleaned.copy()
Portfolio_cleaned.shape[0]/365


df_VOLUNTARY_Account_Number = pd.DataFrame(Portfolio_cleaned[Portfolio_cleaned['ATTRITION_FLAG'] == 'VOLUNTARY'].Account_Number)
df_VOLUNTARY_Account_Number.shape
df_VOLUNTARY_Account_Number = pd.DataFrame(df_VOLUNTARY_Account_Number)


len(df_VOLUNTARY_Account_Number)


#### Remove duplicated accounts from df

VOLUNTARY_Account_Number_dup = df_VOLUNTARY_Account_Number[df_VOLUNTARY_Account_Number.duplicated(keep=False)]
VOLUNTARY_Account_Number_dup.shape


Portfolio_cleaned.shape
Portfolio_cleaned.shape[0]/365
Portfolio_cleaned = Portfolio_cleaned[~Portfolio_cleaned['Account_Number'].isin(list(VOLUNTARY_Account_Number_dup['Account_Number']))]
Portfolio_cleaned.shape[0]/365

unique_accounts_count = len(Portfolio_cleaned['Account_Number'].unique())
unique_accounts_count
df = Portfolio_cleaned


len(df_VOLUNTARY_Account_Number['Account_Number'].unique())
unique_accounts_count
len(df_VOLUNTARY_Account_Number['Account_Number'].unique())/unique_accounts_count

### Create List of Open Account Numbers to Resample Timespan

df_OPEN = df.loc[~df['Account_Number'].isin(df_VOLUNTARY_Account_Number['Account_Number'].unique())]

df_OPEN.isnull().sum()

unique_OPEN = df_OPEN['Account_Number'].unique()
len(unique_OPEN)
unique_Account_Number = df['Account_Number'].unique()
len(unique_Account_Number)
unique_VOLUNTARY = df_VOLUNTARY_Account_Number['Account_Number'].unique()
len(unique_VOLUNTARY)

#assert len(unique_Account_Number) == (len(unique_OPEN) + len(unique_VOLUNTARY))

type(df_OPEN)
df.shape
df_OPEN.shape
df_OPEN.shape[0]/365


### Get range of MISDATE where df_OPEN

extract earliest possible start dates for each account

# Check min date
min_dates = df_OPEN.groupby('Account_Number')['MISDATE'].min()
# Check max date
max_dates = df_OPEN.groupby('Account_Number')['MISDATE'].max()

to_dates = max_dates
from_dates = min_dates + pd.DateOffset(days=130)

dates = pd.DataFrame(to_dates); dates.shape; dates.columns; 
dates.reset_index(inplace=True); #dates
dates.rename(columns={"MISDATE": "to_dates"}, inplace=True)

dates['from_dates'] = pd.DataFrame(from_dates.values)
dates.shape
dates = dates[dates['to_dates'] == '2017-09-30']
dates.shape
df_OPEN.shape[0]/365
len(df_OPEN['Account_Number'].unique())

dates.isnull().sum()

### Create a Random Selection of `start_date` for Sample

dates.columns
dates.shape
dates.to_dates.value_counts()

from random import randint
import datetime
datetime.timedelta(randint(1,234))
len(dates['to_dates'])
from random import randint
import datetime
dates['randomint']=""
dates['randomint'] = np.random.choice(range(1,234),len(dates),replace=True)


dates['random_end_dt']=dates['to_dates']-pd.to_timedelta(dates['randomint'],unit='d')
dates.head(2)
dates.shape

df_OPEN1 = pd.merge(df_OPEN, dates, how='left',on='Account_Number')

del df_OPEN1['to_dates']
del df_OPEN1['from_dates']
df_OPEN1.shape
df_OPEN1.head(3)

df_OPEN2 = df_OPEN1[df_OPEN1['MISDATE'] <= df_OPEN1['random_end_dt']]
df_OPEN2.head(2)

del df_OPEN2['randomint']

df_OPEN2 = df_OPEN2.reset_index(drop=True)
df_OPEN2 = df_OPEN2.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False])
df_OPEN2 = df_OPEN2.reset_index(drop=True)
df_OPEN2 = df_OPEN2.groupby('Account_Number').head(130)
df_OPEN2 = df_OPEN2.reset_index(drop=True)
df_OPEN2.shape
df_OPEN2.head(2)

len(df_OPEN2)/130

len(df_OPEN2[df_OPEN2['Account_Number']=="MAFF_ACCT_0000266356"])
len(df_OPEN2)/130

del df_OPEN2['random_end_dt']

output = open('Open_Sampling_Ready1.pkl', 'wb')
pickle.dump(df_OPEN2, output)
output.close()

df.shape
df_OPEN2.shape
df_VOLUNTARY_Account_Number.shape
df_VOLUNTARY_Account_Number.head(2)

df_VOLUNTARY = df[df['Account_Number'].isin(list(df_VOLUNTARY_Account_Number['Account_Number']))]

df_VOLUNTARY.shape
df_VOLUNTARY.Account_Number.nunique()
df_VOLUNTARY_Account_Number.values
len(df_VOLUNTARY)/365

df_VOLUNTARY = df_VOLUNTARY.reset_index(drop=True)
df_VOLUNTARY = df_VOLUNTARY.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False])
df_VOLUNTARY = df_VOLUNTARY.reset_index(drop=True)
df_VOLUNTARY = df_VOLUNTARY.groupby('Account_Number').head(130)
df_VOLUNTARY = df_VOLUNTARY.reset_index(drop=True)

len(df_VOLUNTARY)/130
df_VOLUNTARY.isnull().sum()

totalframe = pd.concat([df_VOLUNTARY,df_OPEN2], axis=0)
totalframe.shape
totalframe = pd.DataFrame(totalframe)

totalframe.isnull().sum()

#Check to see our data is divisable by 130 timesteps
len(totalframe)/130

#Write out sample
output = open('All_Sampling_Ready1.pkl', 'wb')
pickle.dump(totalframe, output)
output.close()

del df_VOLUNTARY
del df_OPEN2
totalframe.shape
totalframe.head(1)

Statements_cleaned = pd.read_csv('Statements_cleaned.csv',sep=',',encoding='utf-8')
del Statements_cleaned['Unnamed: 0']
Statements_cleaned.shape
Statements_cleaned.head(2)
Statements_cleaned.dtypes
#list(Statements_cleaned.columns.values)

CRM_cleaned = pd.read_csv('CRM_cleaned.csv',sep=',',encoding='utf-8')
del CRM_cleaned['Unnamed: 0']
CRM_cleaned.shape
CRM_cleaned.head(2)
CRM_cleaned.dtypes
#list(CRM_cleaned.columns.values)

Transaction_DailyAgg_cleaned = pd.read_csv('Transactions_DailyAgg_cleaned.csv',sep=',',encoding='utf-8')
del Transaction_DailyAgg_cleaned['Unnamed: 0']
Transaction_DailyAgg_cleaned.shape
Transaction_DailyAgg_cleaned.head(2)
Transaction_DailyAgg_cleaned.dtypes

Transaction_DailyAgg_cleaned = Transaction_DailyAgg_cleaned.dropna(how='any',axis=0)

Statements_cleaned.isnull().sum()
CRM_cleaned.isnull().sum()
Transaction_DailyAgg_cleaned.isnull().sum()

#### Format Select Variables

CRM_cleaned['Account_Number'] = CRM_cleaned['Account_Number'].astype(str)
CRM_cleaned['Created_On'] = CRM_cleaned['Created_On'].astype('datetime64[ns]')

del Portfolio_cleaned

#### Join Key Incident Data from CRM to Portfolio Dataframe.

CRM_cleaned.shape
threshold = 200
counts_type = CRM_cleaned['Incident_Type'].value_counts()
counts_cat = CRM_cleaned['Incident_Category'].value_counts()

CRM_cleaned = CRM_cleaned.loc[CRM_cleaned['Incident_Type'].isin(counts_type[counts_type > threshold].index), :]
CRM_cleaned = CRM_cleaned.loc[CRM_cleaned['Incident_Category'].isin(counts_cat[counts_cat > threshold].index), :]
CRM_cleaned = CRM_cleaned.reset_index(drop=True)
CRM_cleaned.shape


#### Truncate CRM Daily Incident Data to 5 Most Recent Per Day

# Truncate to 5 incidents per account per date
CRM_cleaned = CRM_cleaned.sort_values(['Account_Number','Created_On'], ascending=[True, True])
CRM_cleaned = CRM_cleaned.reset_index(drop=True)

Account_Incident_Counts = CRM_cleaned.groupby(['Account_Number', 'Created_On']).size()
Account_Incident_Counts = Account_Incident_Counts.reset_index()
Account_Incident_Counts.columns = ['Account_Number', 'Created_On', 'Incident_Num']
Account_Incident_Counts = pd.DataFrame(Account_Incident_Counts)
Account_Incident_Counts.head(1)

CRM_cleaned.shape
CRM_cleaned2 = CRM_cleaned.sort_values(['Account_Number','Created_On'], ascending=[True, True])
CRM_cleaned2 = CRM_cleaned2.groupby(['Account_Number','Created_On']).head(5)                                                         
CRM_cleaned2 = CRM_cleaned2.reset_index(drop=True)


Account_Incident_Counts.isnull().sum()

#### Concatanate sequences of the category and type of the day's incidents for that customer

# Aggregate longer sequences of multiple incidents per day into its own variable.
CRM_cleaned_seq = CRM_cleaned.groupby(['Account_Number', 'Created_On'])['Incident_Category'].agg(lambda col: ' _ '.join(col))
CRM_cleaned_seq = pd.DataFrame(CRM_cleaned_seq)
CRM_cleaned_seq = CRM_cleaned_seq.reset_index(drop=True)
CRM_cleaned_seq.columns = ['Incident_Cat_Sequence_Daily']
CRM_cleaned_seq.shape

# Aggregate longer sequences of multiple incidents per day into its own variable.
CRM_cleaned_seq2 = CRM_cleaned.groupby(['Account_Number', 'Created_On'])['Incident_Type'].agg(lambda col: ' _ '.join(col))
CRM_cleaned_seq2 = CRM_cleaned_seq2.reset_index(drop=True)
CRM_cleaned_seq2 = pd.DataFrame(CRM_cleaned_seq2)
CRM_cleaned_seq2.columns = ['Incident_Type_Sequence_Daily']
CRM_cleaned_seq2.shape

#### Join the new variables to the data frame.

CRM_cleaned_seq = CRM_cleaned_seq.join(CRM_cleaned_seq2)
CRM_cleaned_seq.head(4)

CRM_cleaned = CRM_cleaned.sort_values(['Account_Number','Created_On'], ascending=[True, True])
CRM_Index = CRM_cleaned.groupby(['Account_Number', 'Created_On']).head(1)
CRM_Index = pd.DataFrame(CRM_Index)
CRM_Index = CRM_Index.reset_index(drop=True)
CRM_Index.shape

CRM_Joined = CRM_Index.join(CRM_cleaned_seq)
CRM_Joined.head(4)

#### Merge CRM Incident Daily Incidents and Sequences to the Working File

Merged2 = pd.merge(totalframe, CRM_Joined,  how='left', 
                  left_on=['Account_Number','MISDATE'], 
                  right_on = ['Account_Number','Created_On'])
del Merged2['Created_On']

del totalframe

Merged2['Incident_Category'].fillna('None', inplace=True)
Merged2['Incident_Type'].fillna('None', inplace=True)
Merged2['Incident_Cat_Sequence_Daily'].fillna('None', inplace=True)
Merged2['Incident_Type_Sequence_Daily'].fillna('None', inplace=True)
Merged2['Incident_Category']=Merged2['Incident_Category'].replace(-np.nan, 'None').replace(np.nan, 'None')
Merged2['Incident_Type']=Merged2['Incident_Type'].replace(-np.nan, 'None').replace(np.nan, 'None')
Merged2['Incident_Cat_Sequence_Daily']=Merged2['Incident_Cat_Sequence_Daily'].replace(-np.nan, 'None').replace(np.nan, 'None')
Merged2['Incident_Type_Sequence_Daily']=Merged2['Incident_Type_Sequence_Daily'].replace(-np.nan, 'None').replace(np.nan,'None')

Merged2['Incident_Category'] = Merged2['Incident_Category'].astype('str')
Merged2['Incident_Type'] = Merged2['Incident_Type'].astype('str')
Merged2['Incident_Cat_Sequence_Daily'] = Merged2['Incident_Cat_Sequence_Daily'].astype('str')
Merged2['Incident_Type_Sequence_Daily'] = Merged2['Incident_Type_Sequence_Daily'].astype('str')

Merged2 = pd.merge(Merged2, Account_Incident_Counts,  how='left', 
                  left_on=['Account_Number','MISDATE'], 
                  right_on = ['Account_Number','Created_On'])
del Merged2['Created_On']

Merged2['Incident_Num'].fillna(0, inplace=True)            
Merged2['Incident_Num'] =Merged2['Incident_Num'].replace(np.nan,0).replace(np.inf,0) 

Merged2.isnull().sum()

Merged2.columns
Merged2.head(2)

Merged2.shape

output = open('Intermediate_preprocessing2.pkl', 'wb')
pickle.dump(Merged2, output)
output.close()


### Create Select One-Hot Encodings for Select Variables

#APR Rate 
Merged2['APR_Rate'].unique()
Merged2['APR_Rate_Low'] = np.where(Merged2['APR_Rate']=='LOW', '1', '0')
Merged2['APR_Rate_Med'] = np.where(Merged2['APR_Rate']=='MEDIUM', '1', '0')
Merged2['APR_Rate_High'] = np.where(Merged2['APR_Rate']=='HIGH', '1', '0')
del Merged2['APR_Rate']

Merged2['APR_Rate_Low'] = Merged2['APR_Rate_Low'].replace(np.nan, 0).replace(np.inf, 0).astype(float)
Merged2['APR_Rate_Med'] = Merged2['APR_Rate_Med'].replace(np.nan, 0).replace(np.inf, 0).astype(float)
Merged2['APR_Rate_High'] = Merged2['APR_Rate_High'].replace(np.nan, 0).replace(np.inf, 0).astype(float)

Merged2['APR_Rate_Low'] = Merged2['APR_Rate_Low'].astype(int)
Merged2['APR_Rate_Med'] = Merged2['APR_Rate_Med'].astype(int)
Merged2['APR_Rate_High'] = Merged2['APR_Rate_High'].astype(int)


#Product Category 
Merged2['Product_Category'].unique()
Merged2['Prod_Cat_Voyager'] = np.where(Merged2['Product_Category']=='VOYAGER', '1', '0')
Merged2['Prod_Cat_Cashback'] = np.where(Merged2['Product_Category']=='CASHBACK', '1', '0')
del Merged2['Product_Category']

Merged2['Prod_Cat_Voyager'] = Merged2['Prod_Cat_Voyager'].replace(np.nan, 0).replace(np.inf, 0).astype(float)
Merged2['Prod_Cat_Cashback'] = Merged2['Prod_Cat_Cashback'].replace(np.nan, 0).replace(np.inf, 0).astype(float)

Merged2['Prod_Cat_Voyager'] = Merged2['Prod_Cat_Voyager'].astype(int)
Merged2['Prod_Cat_Cashback'] = Merged2['Prod_Cat_Cashback'].astype(int)

Merged2_backup = Merged2
Merged2['Nationality_Group'] = Merged2['Nationality_Group'].replace('01. INDIA', 'INDIA')
Merged2['Nationality_Group'] = Merged2['Nationality_Group'].replace('02. PAKISTAN', 'PAKISTAN')
Merged2['Nationality_Group'] = Merged2['Nationality_Group'].replace('03. PHILLIPPINES', 'PHILLIPPINES')
Merged2['Nationality_Group'] = Merged2['Nationality_Group'].replace('04. U.A.E.', 'U.A.E.')
Merged2['Nationality_Group'] = Merged2['Nationality_Group'].replace('05. ARAB', 'ARAB')
Merged2['Nationality_Group'] = Merged2['Nationality_Group'].replace('06. OTHERS', 'OTHERS')

del Merged2_backup

Merged2= pd.get_dummies(Merged2, columns=['Nationality_Group'], prefix=['Country'])


#Format new int variables as int
Merged2['Country_INDIA'] = Merged2['Country_INDIA'].replace(np.nan, 0).astype(int)
Merged2['Country_PAKISTAN'] = Merged2['Country_PAKISTAN'].replace(np.nan, 0).astype(int)
Merged2['Country_PHILLIPPINES'] = Merged2['Country_PHILLIPPINES'].replace(np.nan, 0).astype(int)
Merged2['Country_U.A.E.'] = Merged2['Country_U.A.E.'].replace(np.nan, 0).astype(int)
Merged2['Country_ARAB'] = Merged2['Country_ARAB'].replace(np.nan, 0).astype(int)
Merged2['Country_OTHERS'] = Merged2['Country_OTHERS'].replace(np.nan, 0).astype(int)



### Derived Features

Merged2['ACTUAL_CURRENT_BALANCE'] = Merged2['ACTUAL_CURRENT_BALANCE'].replace(np.inf, 1).replace(np.nan, 0).astype(float)
Merged2['Credit_Limit'] = Merged2['Credit_Limit'].replace(np.inf, 1).replace(np.nan, 0).astype(float)

Merged2['Credit_Limit'] = Merged2['Credit_Limit'].round().astype(int)
Merged2['ACTUAL_CURRENT_BALANCE'] = Merged2['ACTUAL_CURRENT_BALANCE'].round().astype(int)

# Percent Credit Used
Merged2['Perc_Credit_Limit_Used'] = (Merged2['ACTUAL_CURRENT_BALANCE'])/Merged2['Credit_Limit']
Merged2['Perc_Credit_Limit_Used'] = Merged2['Perc_Credit_Limit_Used'].round(2)
Merged2['Perc_Credit_Limit_Used'] = Merged2['Perc_Credit_Limit_Used'].replace(np.inf, 1).replace(np.nan, 0).astype(float)


del Merged2['ACCOUNT_CREATION_DATE']
del Merged2['Last_Statement_Date']
del Merged2['LAST_ACTIVITY_DATE']
del Merged2['Next_Statement_Date']

### Write out intermediate file

#Merged2.to_csv('Merged_cleaned.csv',sep=',',encoding='utf-8')
output = open('Merged_cleaned2.pkl', 'wb')
pickle.dump(Merged2, output)
output.close()

#Merged2.to_csv('Merged_cleaned.csv',sep=',',encoding='utf-8')
df1 = Merged2


del Merged2

### Import and Merge Statements Table Data of Interest

df2 = pd.read_csv('Statements_cleaned.csv',sep=',',encoding='utf-8')
del df2['Unnamed: 0']
del df2['STATEMENT_FROM_DATE']
del df2['STATEMENT_TO_DATE']
df2.head(2)

df2['STATEMENT_DATE'] = df2['STATEMENT_DATE'].astype('datetime64[ns]')
df3 = pd.merge(df1,df2,how='left', 
                  left_on=['Account_Number','MISDATE'], 
                  right_on = ['Account_Number','STATEMENT_DATE'])
del df3['STATEMENT_DATE'] 
del df3['STATEMENT_BALANCE']
df3['POINTS_REDEEMED']=df3['POINTS_REDEEMED'].replace(np.nan, 0).replace(np.nan, 0)
df3['POINTS_BALANCE_CLOSING']=df3['POINTS_BALANCE_CLOSING'].replace(-np.nan, 0).replace(np.nan, 0)
df3['POINTS_EARNED']=df3['POINTS_EARNED'].replace(-np.nan, 0).replace(np.nan, 0)

df3['POINTS_REDEEMED'].fillna(0, inplace=True)
df3['POINTS_BALANCE_CLOSING'].fillna(0, inplace=True)
df3['POINTS_EARNED'].fillna(0, inplace=True)

df3['POINTS_REDEEMED'] = df3['POINTS_REDEEMED'].astype('int')
df3['POINTS_BALANCE_CLOSING']= df3['POINTS_BALANCE_CLOSING'].astype('int')
df3['POINTS_EARNED']= df3['POINTS_EARNED'].astype('int')
df3.shape

del df2

os.chdir('D://MAFData//')
os.getcwd()

with open('Transactions_cleaned.pkl', 'rb') as f:
    Transactions_DailyAgg_cleaned = pickle.load(f)


df1['Account_Number'] = df1['Account_Number'].astype('str')
df1['Account_Number'] = df1['Account_Number'].astype('str')

Transactions_DailyAgg_cleaned['Purchase_Date'] = Transactions_DailyAgg_cleaned['Purchase_Date'].astype('datetime64[ns]')
df1['MISDATE'] = df1['MISDATE'].astype('datetime64[ns]')

Transactions_DailyAgg_cleaned = Transactions_DailyAgg_cleaned[pd.notnull(Transactions_DailyAgg_cleaned['Account_Number'])] 
Transactions_DailyAgg_cleaned = Transactions_DailyAgg_cleaned[pd.notnull(Transactions_DailyAgg_cleaned['Purchase_Date'])] 

df1.tail(1)
Transactions_DailyAgg_cleaned.tail(1)


df3 = pd.merge(df3,Transactions_DailyAgg_cleaned,how='left', 
                  left_on=['Account_Number','MISDATE'], 
                  right_on = ['Account_Number','Purchase_Date'])
del df3['Purchase_Date']

del df1

df4 = df3
del df3

#Merged2.to_csv('Merged_cleaned.csv',sep=',',encoding='utf-8')
output = open('Merged_cleaned2.pkl', 'wb')
pickle.dump(df4, output)
output.close()

### Replace remaining nan and infinities with '0'

df4['Billing_Amount_Sum_Spend_ESS']=df4['Billing_Amount_Sum_Spend_ESS'].replace(-np.nan, 0).replace(np.nan, 0)
df4['Billing_Amount_Sum_Spend_NON_ESS']=df4['Billing_Amount_Sum_Spend_NON_ESS'].replace(-np.nan, 0).replace(np.nan, 0)
df4['Billing_Amount_Spend_CASH']=df4['Billing_Amount_Spend_CASH'].replace(np.nan, 0).replace(np.nan, 0)
df4['Billing_Amount_Violation_Fee']=df4['Billing_Amount_Violation_Fee'].replace(-np.nan, 0).replace(np.nan, 0)
df4['Transaction_Count_Payment']=df4['Transaction_Count_Payment'].replace(-np.nan, 0).replace(np.nan, 0)

df4['Tranasaction_Count_Spend_ESS']=df4['Tranasaction_Count_Spend_ESS'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Tranasaction_Count_Spend_NON_ESS']=df4['Tranasaction_Count_Spend_NON_ESS'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Tranasaction_Count_Spend_CASH']=df4['Tranasaction_Count_Spend_CASH'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Transaction_Count_Foreclosure']=df4['Transaction_Count_Foreclosure'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Tranasaction_Count_BT']=df4['Tranasaction_Count_BT'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Tranasaction_Count_LOP']=df4['Tranasaction_Count_LOP'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Tranasaction_Count_EPP']=df4['Tranasaction_Count_EPP'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Tranasaction_Count_Violation_Fee']=df4['Tranasaction_Count_Violation_Fee'].replace(np.nan, 0).replace(np.inf, 0).astype(int)

df4.head(1)

df4['Incident_Num']=df4['Incident_Num'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['APR_Rate_Low']=df4['APR_Rate_Low'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['APR_Rate_Med']=df4['APR_Rate_Med'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['APR_Rate_High']=df4['APR_Rate_High'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Prod_Cat_Voyager']=df4['Prod_Cat_Voyager'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Prod_Cat_Cashback']=df4['Prod_Cat_Cashback'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Days_On_Book']=df4['Days_On_Book'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Days_From_Statement']=df4['Days_From_Statement'].replace(np.nan, 0).replace(np.inf, 0).astype(int)
df4['Days_from_Last_Activity']=df4['Days_from_Last_Activity'].replace(np.nan, 0).replace(np.inf, 0).astype(int)

df4.dtypes

### Add Timestep Variable

df4 = df4.sort_values(['Account_Number', 'MISDATE'], ascending=[True, False])
df4 = df4.reset_index(drop=True)
df4['Timestep'] = df4.groupby(['Account_Number']).cumcount()+1
df4.shape
df4.head(4)
df4.tail(4)

### Save Local Copy


output = open('Merged_cleaned2.pkl', 'wb')
pickle.dump(df4, output)
output.close()

#df4.to_csv('Merged_cleaned2.csv',sep=',',encoding='utf-8')

output = open('Merged_cleaned2_baknov2.pkl', 'wb')
pickle.dump(df4, output)
output.close()
